Add plan tags as rework tags when loading changes from the database.
This is necessary if, for example, a change has been reworked, but the
reworked version is not deployed, only the previous version. In that case,
Sqitch did not realize it was dealing with a reworked version, so did not look
for a file named `$change@$tag.sql`. So for each change, also look in the plan
to see if the change has been reworked, and if it has, associate all tags as
rework tags up to the reworking change. This should allow the proper deploy,
verify, and revert scripts to be found.

Resolves #166.