Use symlinkeds for Oracle scripts with special characters.
Instead of just relying on `ESCCHAR`, because that's only available on
SQL*Plus 11.1 or higher, and we claim to support 8.4.0! So we instead create a
temporary directory, which is deleted on exit, and either symlink to that
directory, or copy the contents of the file (on Windows).

Should be a better solution to issue #165.