Fix function lookup errors with Text::ANSIColor.
Not sure why it suddenly wanted `color` to be treated as an indirect method
call, but since it also conflicts with an attribute on ItemFormatter, alias it
to `encolor` and call that with parentheses, so that it is always called as a
function. Use parens with `colorvalid`, too, just to be safe.

Resolves #168.