Don't exit with 1 on deploy to up-to-date db.
Instead, just send the message that the database is up-to-date to info() and
return success. This has been a long-standing request. I audited all the
places where we return an exit value of 1, and this is the only one anyone has
complained about, I believe. And on reflection, I think it is right that it
should not be an error.

Resolves #147.