Add the `default_target` attribute to Command.
To be used by commands that don't require access to an engine, but do need to
know about the plan, or file locations. Examples include `init` and `add`.

Also expand the documentation for `parse_args()` to better record how it
creates targets, especially the default target, which in this case *must* have
an associated engine.