Deprecare core.$engine in favor of engine.$engine.
Makes it easier to tell when there is old stuff, so we can emit a warning and
suggest that users run `sqitch engine update-config` to fix it. It also makes
it more in line with the precedent for targets, and, in the future, VCSes.

This may be an annoying change for some folks, but just running `sqitch engine
update-config` once per project should just take care of it.