Limit merging of deprecated URI parts.
Should only merge them if the URI came from `core.target` or
`core.$engine.target`. This version allows it if the URI came from
`engine.$engine.target`, but if that key is present, none of the deprecated
configs should be, and it's proabably sensible to go ahead an merge
command-line options.

So the logic now has two new paths:

* If a URI is passed in the name, it was specified by an option or in a
  command-line argument, so we just take it and run, no merging or anything
  else.
* Add a lookup for `core.target`, which was missing.