Don't delete old core config.
Instead, let it live on, so that it exists for those who still have older
versions. Downside is that it may not be kept in sync, but I think this
is the less painful approach.

As a result, don't complain if there is a `core.$engine` config *and* a an
`engine.$engine` config.

Also fix l10n to be on single lines, so that parser can find them.