Add registry updating to the engine.
It will automatically run an upgrade on `deploy`. This commit includes the
first upgrade, to 1.0, which is the addition of the `releases` table.

Had to change the Oracle engine to always `DEFINE` the `registry` variable,
like Postgres does. Also tweaked the MySQL engine to use the new
`run_upgrade()` method for the initialization script, and to catch new text
for a missing table error.