Add the `script_hash` column to the `changes` table.
The SHA-1 hash of the deploy script will be stored here. However, since this
value is not available for all projects during an upgrade, upgraded registries
will instead get a copy of the change ID. A forthcoming commit will add code
to update this value on deploy. And in the future, this value will be used to
help reconcile and merge differences between the plans and registries.

In order to facilitate testing upgrades, an older version of the schema for
each engine registry may now be found in `t/lib/upgradable_registries/`. Even
future engines will need to include this file, with the `script_hash` column
removed, as well as an upgrade script to add it in a
`lib/App/Sqitch/Engine/Upgrade/` script.