Update script hashes.
On deploy, if the latest change ID is the same as its script_hash, update the
script_hashes for all changes in the plan. Items in the database but not the
plan will have their hashes set to NULL. Changes in the plan but not the
database of course will not be added to the database until they're deployed.

Last bit of infrastructure for script hashes as required by the simple merging
proposed in issue #200.