Get Firebird upgrades working properly.
Need to detect when a database file is missing when detecting missing tables.
Also, need to apply the updates to the Sqitch database, not the target
database. And finally, we need to insert the script hash into the changes table,
and omit the `NULL` specification for the `script_hash` column.

The updating of the script hash isn't working on Firebird, for some reason.
Will fix soon.