Be smarter about upgrading the registry.
It was failing on deploy because `current_state` would die on an older
registry with no `script_hash` column. So have that method detect the missing
column and try again without the column. This also allows the `status` command
to work on older registries, since it does not need the script_hash column.