Add SQITCH_PASSWORD environment variable.
This is to unify the handling of environment variables by engines that
support them.

Though on reflection, maybe we don't want this, since not all engines make use
of environment variables, and the use of environment variables for passwords
isn't a great idea to begin with. Will need to ponder it some more.

So I will push it to a branch for now, then review once I've written
`sqitch-passwords` (see #202).