Reduce minimum required MySQL to 5.5.0.
The only feature lost is fractional seconds on `DATETIME` columns. Since the
proper ordering of events is so essential to the way Sqitch works, this commit
introduces a new DBI engine method, `_prepare_to_log()` which the MySQL engine
uses to sleep until the time has ticked over to the next second. This allows
ordering to remain correct at the expense of slower deploys and reverts -- no
more than one change can be deployed or reverted per second.

As a side effect, Sqitch users who start out on a version of MySQL lower than
5.6.4 who then upgrade will have deploys fail. I'll work on an upgrade script
to add a precision to the `DATETIME` columns for those folks.

Closes #218.