Fix event ordering test failure on MySQL 5.5.
There's a test that applies a regular expresion against projects, matching
multiple. On MySQL 5.5, we wait for a second to tick over, but only on a
per-project basis. So to make sure the test gets things back in order, pretend
to prepare to insert another "engine" project item before inserting a
"groovy" project item. This way, we're sure the clock will have ticked over
before inserting the "groovy" change.