Consistently set rspace on a note.
Unless the line has no name, which should only be true for blank lines. This
makes it so that newly added plan lines -- changes and tags -- should now
always have a space befor the `#` character that marks the beginning of a
note. A line that lacks the space when a plan is parsed won't get a space
unless the note is changed.