Teach `bundle` how to deal with multiple plans.
It now writes the plan files and scripts for all plans known in the Sqitch
configuration, not just the plan for the default target. If you want to limit
the bundle to specific plans, specify the targets for thos plans when calling
`bundle`. Will extend this shortly to allow plan paths and perhaps engine
names to be passed, as well.

Ref #193.