Add changes to all plans and scripts to all top dirs.
And allow the plans and top dirs to be winnowed down by specifying target,
engine, or plan file names on the command-line. Add the `--change` option
to allow the change name to be disambiguated from those other arguments, if
necessary.

The files generated need not be unique by plan; if they're not, they'll be
created by the templates for the engine represented by the default plan (if
any, as defined by core.engine and --engine), or by whatever target was
happened upon first (generally preferring targets over engines). Subsequent
plans that reference the same script files will not create those script files.

Template choice can still be overridden by the `--template-dir` and
`--template-name` options, in which case the resulting templates will be used
to generate the scripts for for *all* top directoris.

If targets and engines use different top directories but the same plan, the
change will be added only to that plan but the scripts created for all of
those top directories.