Teach the tag command to read/write all plans.
By default, it will list the tags or add a tag to all the plan files in a
project. As with the `add` command, allow the plans to be winnowed down by
specifying target, engine, or plan file names on the command-line. Add the
`--tag` and `--change` option to allow the tag and change names to be
disambiguated from those other arguments, if necessary.