Add the --all option to the add command.
And go back to using the default target to determine which plan to add to. Add
the `add.all` configuration option to allow it to be set everywhere,
especially for projects that want to keep plan tags in sync.

This is exactly the same as the behavior of the tag commaand's `--all` option,
added in 0314ef9. So move the code that does the parsing into the Command
class. Eventually will probably want to merge it into the `parse_args()`
method, probably when we address #222.