Replace unique constraint on changes.script_hash.
In order to allow multitle projects to use a single deploy script, let the
constraint be on `project` and `script_hash`, instead. This requires upgrading
the registry to v.1.1. Note that we're also changing the data type of the
`releases.version` column on MySQL so that its precision is always accurate
for our needs.