Eliminate rounding with sprintf().
It gets screwed up by locale settings, where it can use , instead of . to
represent a decimal point. It can be made to work with `no locale;` on 5.20,
but not earlier versions of Perl, where
[a bug](https://metacpan.org/pod/distribution/perl/pod/perl5200delta.pod#Locale-decimal-point-character-no-longer-leaks-outside-of-use-locale-scope)
prevented that from working. So go back to having the databases handle the
rounding. The sprintf was there because Firebirds' round() function sucks,
so just work around that suckage by using substring() in Firebird.

Resolves #234.