Let Oracle deploy > 1000 changes.
Oracle has a configured limit on the number of items that may appear in an
IN() clause. So break it up into mutliple IN() clauses of no more than 250
items each. Oracle allows 1000 by default, but it's configurable, so let's
just try to play it safe by assuming it could be as low as 200.