Teach engine command about reworked directories.
And reformat the output of `show` to be a bit less flat, so hopefully
easier to read.

I'm not entirely happy with the proliferation of `set-` actions in this
command. I originally borrowed it from `git remote`, but starting to think it
might have been better to just use a --set option like the `add` command uses.
I still might go back to that...