Add all other --set options to init.
Require the new `property_keys` method in the class that applies the
TargetConfigCommand role, and use it to enforce what keys are allowed. Then
just use the properties thus specified when writing the configuration file.

New properties include engine, top_dir, client, plan_file, and registry. Each
will get the value passed via the core option, though use of --set is
preferred. Also added target, which will cause a target section to be added to
the config if the target is not a URI. Target, registry, and client are all
added to the engine section.

The init command also injects the engine into the core --engine option, if
it's not specified, since Target looks there to load the proper engine and
create the proper URI. We need that for defaults. It's a bit hinky, but it
works.

The upshot is that the interface is now almost identical to that of the engine
command.

And simplify how things are written to the configuration file.