Make deployed ID search with offset consistent.
In places where we were searching for changes, offsets triggered by `~` and
`^` were properly being handled. Such was the case in the `revert` command.
But in places where we were just searching for IDs, offsets were ignored. So
fix by adding two new engine methods:

* `find_change_id()` is just like `find_change()`, but returns an ID.
* `change_id_offset_from_id()` is just like `change_offset_from_id()`,
  but returns an ID.

The former is a regular Engine method, but the latter is an abstrct method
that engines must implement. So add it to the App::Sqitch::Role::DBIEngine and
each engine class that overrides it.

Resolves #272.