Read dos-formatted config files correctly on unix
It turns out that an extra `^M` character appears at the end of
configuration values read in from e.g. `sqitch.conf` if the configuration
file uses dos-style line endings and the system reading the file is
Unix-based.  The trailing control character caused files and paths from
`Path::Class` to be mis-represented and thus for various other files to not
be found.  In the particular example where this was found, the configuration
for `top_dir` was: `top_dir = sqitch`, which caused the string
representation of `plan_file` (expected to be called `sqitch/sqitch.plan`)
to be `/sqitch.plan`.  The `^M` caused a carriage return and thus overwrote
the directory portion, `sqitch`, (generated from `top_dir`) to be overwritten
with the directory separator and file name part of `plan_file`.  In other
words the path `sqitch^M/sqitch.plan` was returned as `/sqitch.plan`.

The example input configuration file and test in this commit expose this
issue; the changes in `App::Sqitch::Config` correct the problem by removing
superfluous `^M` characters in the data loaded from file as part of
`Config::GitLike`'s `load_file` method.  This certainly fixes the problem,
however there might be a more appropriate location for the fix.

This patch also fixes two corner use-cases: when the input configuration
file happens to contain dos-style line endings, and when sqitch is used in a
Unix-based virtual machine running on Windows and using directories shared
with the host system.  In particular, the issue addressed here was found
when using sqitch in Debian Jessie guest system while using a shared
directory from a Windows 7 host system.  Hence, one is now able to use
sqitch in the guest system even though the project files hosted on the host
system use dos-style line endings.