configuration: Add the App'Sqitch'Config'get_last_value method.
When a property has multiple values in the config file, but we don't
want to error out just because, one of the sane things to do could be to
get the last value in the config for the property.

An example of such a "non-essential" property could be the pager program
to use, core.pager. If it has multiple values set, we likely don't want
to bail out, but just do a best effort pick by using the last value for
the property.