config: Recognize the core.pager property as the pager program to use for all commands.
Commands like `sqitch log` output ANSI color codes when a capable
terminal is detected. By itself, this output shows fine on most terminal
emulators, but when piped to, say, `less`, the output can include the
escaped color codes instead of colored text, depending on the system.
Now IO::Pager looks for the PAGER environment variable, and the user
could just set that to, say, `less -r` to properly show colors. This
patch just adds support for a sqitch specific pager program, specified
by the core.pager property.

If %ENV does not contain "PAGER", this patch sets it to undef, which
works with IO::Pager, although it is not technically the same (!exists
$ENV{PAGER} is different than !$ENV{PAGER}). If we _really_ need to
preserve the existence/non-existence of keys while localizing, we can do
something like

    goto JMP unless exists $ENV{PAGER};
    local $ENV{PAGER} = $config->get_last_value(key => "core.pager");
    JMP:
    ...

But fortunately we don't need that yet.