Improve PostgreSQL version detection.
We only do it when deploying the registry, so it's cheap to just fetch it from
the server as an integer rather than parse the version string from the output
of `psql --version`. It's cleaner, and it avoids issues where the client
version is higher than the server, which could result in sending unsupported
SQL to older versions of Postgres.