Tweak revert and verify change searches.
If change "foo" is reworked, and one says `revert --to foo`, does one want to
revert to the the most recent version of "foo", or to the earlist version of
"foo"? Sqitch has traditionally assumed the latter; this change implements
the former, and does the same for the `--to` option to `verify`.

This might be a bit controversial, however, because some folks might decide
the latter is the more expected case (especially since, if you really want the
latest version you can say `revert --to foo@HEAD`). So this change may well
get reversed pending consensus from the community.

Meanwhile, I also noticed that changes specified as "HEAD" or "ROOT" which are
supposed to be aliases for "@HEAD" and "@ROOT" since v0.991, does not actually
work. This change fixes this issue, as well. This fix likely won't be
reverted, because frankly it has annoyed me for quite some time.