Die on ambiguous change argument.
Alternate fix to #312. Applies to `revert` and `verify`. When a change is
specified by name only and the change appears in the database more than once
(becuase it has been reworked), instead of picking the first or last instance
(as was done in 806d633), raise an error, instead. When doing so, emit a
message telling the user that the change was ambiguous, and list all the
alternates with proper tag-qualification. This is done in change_id_for(),
which is now documented to rais an error on ambiguous changes. The emitted
name and tag are looked up in the database via the existing
`name_for_change_id` engine method. It looks like this method wasn't actually
being used anywhere, so also modify it to append `@HEAD` when appropriate.

Naturally, revert most of the work that went into 806d633a. Also fix a bug
in `revert` where it returned an error for duplicate change names in the
plan when it found nothing in the database. Done by changing `$plan->get`
to `$plan->find`.