Convert Postgres float to text.
On 32-bit Perls, I'm guessing, selecting a REAL column causes the DBI to read
the value in and load it into a float. This has the effect of truncating the
value, at least some of the time. To whit, issue #343 reports that it was
getting back `1` instead of `1.1`. So fix by casting to text before returning
it to the DBI. Resolves #343.