Test vertica connection in before_script.
Which is where we test Exasol. Also comment-out the Exasol DBI connection
test, and have the test output show directives. It also no longer runs in
parallel, because then the directives are not shown. So we have to run them
serially. Not likely a big deal given the overall time to build the
environment for testing.

Oh, and install Template Toolkit, so we get a bit more coverage for the add
command.