Switch Travis CI to Firebird 3.
Firebird 2.5 stpped working sometime in the last few months (a permissions
issue or something). And after a full day of experimenting with various fixes,
trying a Docker image, trying 3.0 and having it fail and all sorts of things
and back and forth again, I finally got 3.0 working by starting it manually.

While at it, show all comments and diagnostics for all tests, and properly
turn multiline text passed to `skip_all` to be quoted as diagnostic text (more
or less). Also just emit diagnostics for a failure to drop a database, since
warnings are fatal.