Adjust snowflake psql execution output.
Although running scripts with `--option quiet=true`, the output still seemed
to include a newline. So capture the output instead, and only emit it if
there is more than whitespace in the output. This works because Sqitch::capture
captures SDTOUT but not STDERR, so no worry of sending errors to STDOUT.

Also, adjust the snowsql verbosity options for _capture, _probe, and _spool to
be sure that they are properly collecting CSV output, and add run_verify so it
can be sure to be verbose when it should be.

Update tests for those changes, as well as the recent change to append
the snowflakecomputing.com domain to the URI.