Just emit output of capture.
For some reason, snowsql emits an extra newline when run directly,
resulting in output that looks like this:

    > sqitch verify
    Verifying flipr_test
      * appschema ..
    ok
      * users ......
    ok

This is super annoying. By capturing the output and just emitting it
ourselves, we prevent snowsql from emitting the extra newline, so the
output looks the way we want:

    > sqitch verify
    Verifying flipr_test
      * appschema .. ok
      * users ...... ok

I had thought that we needed to remove the wayward newline, so made
that happen in 99f4676. But it turns out it's not necessary. So just
emit it ourselves.

Also, add `USE WAREHOUSE &warehouse;` to the templates. It won't
always be needed, but including it is harmless, since Sqitch itself
will already be using that warehouse, and it will serve as a reminder
to the developer to use a warehouse when they need it, with a
reasonable default that should just work.