Add _regex_expr() to Role::DBIEngine and Snowflake.
Because Snowflake implicitly anchors regular expressions to both ends
of the string, we can't use the REGEXP infix operator. The workaround
is to use `regexp_subst() IS NOT NULL`, which of course isn't an infix
operator.  So generalize the handling of regular expresssions for all
engines by adding the _regex_expr() method, with a default in
DBIEngine and an implementation using the Snowflake hack. This finally
gets all the Snowflake engine tests passing.

Whiel at it, add comments documenting known bugs and workarounds, with
links to the appropriate snowflake community questions and bug reports.