Add SQITCH_ORIG_ environment variables.
I realized that `SQITCH_USER_NAME` and `SQITCH_USER_EMAIL`, on their surface,
would be expected to take precedence over values store in the configuration.
So I've changed the to do so, then added `SQITCH_ORIG_NAME` and
`SQITCH_ORIG_EMAIL` to serve as fallbacks for when configuration is *not*
set. They're called "ORIG" because the expectation is that they get set by an
originating host that executes Sqitch on some other host. The canonical
example is the simple shell script to run Sqitch on Docker:

  https://github.com/sqitchers/docker-sqitch/blob/master/docker-sqitch.sh