Always fall back on the system username.
When authenticating to an engine, that is. This is not quite complete, though,
as each engine that supports its own username environment variable must get a
little reworking, since previously we were letting it be implicit, but now
must make it explicit. Will try to document them all in
`sqitch-environment.pod` while I'm at it.