Donn't fallback for Postgres user and password values.
We were not comprehensive enough, failing to find data in the service file.
Best just to let libpq do the work, and not try to replicate it. The downside
is that users of the Docker image may be surprised to find that their default
username is `root` (or whatever the docker image username is(), so we'll just
have to document that behavior. Closes #410.