Allow dependency on reworked changes.
I ran into an issue when I tried to deploy a project that depended
on a reworked change from another project:

    Change "func_logger" is ambiguous. Please specify a tag-qualified change:
      * func_logger@v2.7.0
      * func_logger@2.0
    Change Lookup Failed

In truth, the database requires either version; furthemore, updating the plan
to tag-qualify the dependency would cause all the change IDs to fail, and thus
deploys to fail.

So for the specific case of depending on changes from another project, if a
change has been reworked, and therefore matches multiple records in the
database, simply select the first deployed version of the change as the
dependency. The cleanest, simplest way to do so was to add support the
parameter `first` to `change_id_for()` and teach it not to raise an error if
`first` is true, but just return the first change ID.

While at it, correct the documentation for `change_id_for()`.