Quote shell on Win32 and use Win32 constant.
Hopefully the proper fix will eventually be merged into IPC::System::Simple
and this hack will simply let things work on older vesions.

While at it, switch to a constant for Windows, so as to optimize away
Windows paths in the code at compile time -- not to mention ensure
that checking for Windows is consistent. Or more so, anyway. There are
a couple places it happens (far fewer than before), and at least one
spot that also checks for cygwin, which we might eventually want to
generalize elsewhere, too.