Improve engine detection via arguments.
Previously the `core.engine` configuration was required. This was a relic of
the earliest implementation of Sqitch engines, before there were separate
configurations and the ability to specify them via the command-line. It got in
the way, since if you had an engine specified on the command-line -- either by
name or implicitly in a target -- it would unexpectedly report that no
engine was specified.

So remove that requirement, and only return an error constructing a target if
a target option is specified, or if no other target is found on the
command-line.

This also allows for the elimination of the `no_default` parameter to
`parse_args()`, since it never really worked quite as described anyway
-- it *would*, in fact, return the default.

This change also greatly reduces the need for the core `--engine` variable,
which was previously required to work around this issue. Now it's truly only
necessary for the `init` command, for which it's already implemented as a
separate option -- and which doesn't apply to this path for parsing arguments.
So also change the error message hwne no engine can be found so that it no
longer recommends use of the `--engine` option.

Changes to the command tests are instantiate a concrete implementation of
command so that the `command` attribute is set and access to the configuration
for the command works properly.