Return proper default target from parse_args.
The docs say that the default will always have an engine specified, so we
can't return `default_target`, which will fall back on `db:`. Since we've
already constructed a default and control for its error, just return it.

This breaks a number of tests that had expected the `tag` and `add` commands
to use the default target and its plan. So teach those tests to write out plan
file when things are added to a plan by the test, and to reload the plan when
the test has executed code that wrote to the plan and we need to reload it to
see what's in it. This requires messing with Plan object internals a bit to
delete lazily-created attributes.