Test multiple versions of Firebird.
Pull a Docker image instead of relying on apt packages. Hopefully it will hang
less often (not too optimistic). This lets us test 3.0, 2.5 super classic, and
2.5 super server (currently not starting, sadly:
https://github.com/jacobalberty/firebird-docker/issues/33).

The existing tests didn't support configuring a host name, so add support for
a `FIREBIRD_URI` variable and set it when configuring Firebird. When it has a
database name, the test no longer tries to create the database, and won't drop
it, although it will still drop the registry databases it creates. It also
tries to be smarter before dropping databases by telling Firebird to kill all
other connections.

Also stub out stages for testing Snowflake and Oracle. Neither passes, of
course, so add an `allow_failures` config to allow them to fail (and Firebird
super server, too).