Fix test failures with Encode v2.99.
Most likely caused by [this
change](https://github.com/dankogai/p5-encode/pull/135). It happens because
Locale::Message's bound textdomain filter always passes the optional second
argument, if if it never has one. In the past, that worked okay beause Encode
treated `undef` the same as 0, so we still got the default malformed data
handling. But now that it doesn't, we have to explicitly give the bound text
domain filter a malformed data check parameter. So we give it the default,
`FB_DEFAULT`.

Also update the latest version of Exasol 6.1 to test.