Fix config section case sensistivity.
Config::GitLike treats sections case-insensitively, but subsections are
case-sensitive. So if a variable was added for a target named "FooBar", Sqitch
could not find it, becaues the config section was named
`target.FooBar.variables`, but the App::Sqitch::Config treated it as
`target.foobar.variables`. So add the `_skey` function to properly handle the
canonicalization of section keys, and add some tests to make sure it works
properly.

While at it, require Config::GitLike 1.15 on all platforms. It's nearly five
years old, so should be safe to require everywhere, and allows for the removal
of some legacy code and workarounds (some of which were looking for 1.08!).

Finally, fix the display spacing of variable keys and values in the output of
`target show` and `engine show`. No idea why it was looking at the spacing of
values instead of keys, but it resulted in warnings and improperly spaced
values.

Resolves #454.