Warn on missing Postgres table.
Several folks have reported confusion when using Sqitch and watching the
Postgres error log, especially when experimnenting with a Postgres docker
image, which sends errors to the terminal. So add a warning to the error log
immediately after the missing relation error to tell log watchers the Sqitch
is aware of the issue and will next initialize the database. Hopefully this
will reduce confusion and unnecessary time trouble shooting an error that
Sqitch handles. Resolves #314.