Add lock_registry engine method
Called at the start of deploy() and revert(). The default implementation does
nothing, but on MySQL and Postgres, it acquires an adivosry lock not released
till the end of the session. This is to ensure that only one instance of
Sqitch is making changes at a time. Raises an error for the user to tell them
that another instance is at work. Not implemented for the other engines as I
could not quickly find a way to create session-level locks for them. See #579
for details.