Revamp destination locking
Change the name from session locking, and implement most of the logic at the
engine base class level. To do so, introduce two new methods for subclasses to
implement, `try_lock` and `wait_lock`. The former should immediately reeturn
true or false when it attempts to take a lock, the latter should try to take
the same lock but wait indefinitely. The `lock_destination` method first calls
`try_lock` and emits a notice if it fails, telling the user it is waiting,
then calls `wait_lock` and times out after a period. It handles all the
timeout waiting, so the database engines can just do the lock stuff.

Add attributes to track the lock acquisition, so we only try to lock once, and
another for a wait timeout.

Future work here will include configuration for the timeout and testing. It
may be possible to implement the locking in Oracle using `DBMS_LOCK.REQUEST`,
too, but it's PL/SQL, not SQL, so I'm taking a pass for now.