Remove Sys::SigAction and restore Windows support
Instead modify wait_lock() to actually do the waiting, and add a new method,
_timeout() that an engine can optionally use to retry lock acquisition with an
exponential backoff. Modify the Postgres engine to use an async query to try
to acquire a lock and use _timeout to check for the result until it times out.
Have the MySQL engine send the timeout to the `get_lock()` function and let it
handle the timeout.

This eliminates the need to rely on system alarm interrupts, which simply
don't work on Windows, and are pretty fussy in any event. Retrying with
backoff is a much simpler approach, especially when using async queries, and
since it will get the lock in the first call most of the time anyway, should
be reasonably safe to use.