Properly build and test
Previously no tests were running, because we were not cloning the repository.
D'oh! So clone the repo and rework things to properly install all the
dependencies and test them. Done by downloading the cpanfile for the previous
release, installing Windows-specific dependencies, all the recommended modules
in the cpanfile, plus a couple extras for testing. Also install the modules
locally and set PERL5LIB to use them. And finally, fix a test failing test on
Windows in t/options.t that incorrectly compared error messages.