Filter for engine-specific changes
Look for changes to files matching `**/pg.*` before running the Postgres
workflow, and to files matching `**/sqlite.*` before running the SQLite
workflow. Use dorny/paths-filter to do so, since the GitHub action `on` syntax
does not support conditionals like this. Instead, we add a job wherein
dorny/paths-filter sets an output variable to "true" if it finds matching files.
Then in the Postgres or SQLite test job, we check that output variable and skip
the job if there were no changes to those paths. But don't do that work on the
main or develop branches; there just run the database testing jobs and skip
the filtering.

Also, modify the postgres and SQLite setup scripts to take an argument, rather
than read from an environment variable; though continue to support a variable,
because why not. Also have them properly append to the PATH variable, and use
the `$GITHUB_ENV` variable to add environment variables to subsequent steps.

And finally, use the `LIVE_*_REQUIRED` environment variables read DBIEngineTest
to require that the live tests run, rather than make a custom on (had forgotten
about this) feature of DBIEngineTest). Also emit diagnostics for the version of
`psql` used in tests.