Add MySQL testing
And require that all the engine workflows trigger if any file matching "Engine"
is modified, not just the engine name.

Modify DBIEngineTest and t/mysql.t to get the tests passing for all available
versions of MySQL and MariaDB; the locking stuff recently added is a bit fussy
between versions, and we need to wait for async release in more than one place
for some of them, and MySQL timeouts cannot be less than a second. So make the
wait timeout configurable, as well as the async waiting and configure it with
appropriate release function for the version of MySQL or Maria.