Add Oracle testing and remove path filter
Tests Oracle 11g and 18c. Should also test 21c, but connections fail, so that's
commented out for now. With tests running, fix issues with the script hash,
noisy output on variable substitution by SQL*Plus, and the lack of ordering of
collections on Oracle 18. Also test with a URI environment variable rather than
TWO_TASK, and specify the name of an alternate schema (user) via a variable, as
`oe` is in the developer days VM but not the Docker images.

The path filter did not work, it turns out, unless I revamped it to re-export
values and then made the subsequent steps depend on the filter. Which means it
always has to run, whether it would be checked or not. It was just getting
messier than I'd like, so let's just go with the convention of naming branches
appropriately to trigger tests.

And finally, add a test comment for the Vertica VSQL client version.