Use SQITCH_TEST_${engine}_URI for all tests.
Mostly supported by a URI parameter for the engines, and we fall back on the
old URI environment variables. The exception is Postgres, which never used a
URI variable. So refactor t/pg.t to support it, and modify the setup script
and the GitHub workflow to use the URI.

Also tweak some of the ways in which the CLI diagnostic is displayed for
some of the engines, and add init scripts for MySQL and Firebird.