Use workflow includes, expand Oracle docs, move scripts
Switch from customized YAML objects and arrays to the `matrix.includes` feature.
This makes it slightly easier to understand the configurations, and simplifies
referencs to the matrix object.

Expand the Oracle tutorial and test comments to cover the use of the Docker-XE
images, as well as the installation of Instant Client and DBD::Oracle. I have
not fully verified the new configurations, but it should be very close given
what I've learned writing the Oracle testing workflow.

Move all of the configuration scripts from `dev/linux-*` to
`.github/ubuntu/*.sh`. This makes it clearer that they are specifically for
configuring the GitHub workflow Ubuntu images, though they're generally
applicable to Ubuntu systems. It will also line up nicely should we decide to
add database client config scripts for macOS and Windows. This left only the
`Dockerfile` and `dependency_report` files in the `dev` directory, so delete the
former, since the https://github.com/sqitchers/docker-sqitch repo is canonical
for Docker files, and move the latter to `xt` with other release tools.

Change the `xt/release/*.t` tests to always load the modules they depend on,
rather than skip the tests if they can't be loaded. This is to ensure that
release tests are always run prior to release. Also sort the additional spelling
words in `pod-spelling.t` and add a couple of new ones from the expanded Oracle
documentation.

Run `xt/dependency_report` and paste the updated list of build-only dependencies
into `inc/Menlo/Sqitch.pm`.