Add test coverage workflow.
Add a new script, `.github/ubuntu/all-apt-prereqs.sh`, to install all apt
packages at once, and teach each of the engine scripts not to install apt
packages when `SKIP_DEPENDS` is set. Also have the Postgres an SQLite scripts
fall back on a default version to install if one is not passed. Remove the old
`prereqs.sh`. And have the SQLite script install SQLite into `/opt/sqlite` to be
more like the other engine installers, and have it and the Oracle script properly
share the `LD_LIBRARY_PATH` variable.

The new workflow, `.github/workflows/coverage.yml`, sets up all of the services
and clients, as well as modules for reporting coverage to Coveralls, and runs
the tests. Now that it uses a script installed in `local/bin`, teach all the
workflows to cache all of `local`, not just `local/lib`.

Get Oracle 21 tests working, too. Thanks to @gvenzl's help in
gvenzl/oci-oracle-xe#45.

Remove `.travis.yml` and references to it. Travis testing has not worked for a
while, and is no more.

Have `prove` run tests in parallel in 4 processes where it runs multiple tests,
so that builds finish more quickly.

Add build and release badges to README.md, and move most links to the bottom of
the text for better plain text legibility. Most of the GitHub badges don't
handle the Emoji correctly, but hopefully that will be addressed before long.

Finally, add a call to `disconnect` to `t/oracle.t` to prevent a segfault
that Devel::Cover seems to induce in DBD::Oracle (see perl5-dbi/DBD-Oracle#111).